LGD-564: fix unit tests + add unit tests for make_ascending
